refactor(releases): rename components and files to releases rather than bundles
