+++ packages/sanity/src/core/releases/components/ReleasesTable/ReleasesTable.tsx
@@ -4,8 +4,8 @@ import {styled} from 'styled-components'
 
 import {type BundleDocument} from '../../../store/bundles/types'
 import {type BundlesMetadata} from '../../tool/useBundlesMetadata'
-import {BundleHeader, type BundleHeaderProps} from './BundleHeader'
-import {BundleRow} from './BundleRow'
+import {ReleaseHeader, type ReleaseHeaderProps} from './ReleaseHeader'
+import {ReleaseRow} from './ReleaseRow'
 
 const RowStack = styled(Stack)({
   '& > *:not(:first-child)': {
@@ -24,11 +24,11 @@ export interface TableBundle extends BundleDocument {
   documentsMetadata: BundlesMetadata
 }
 
-interface BundlesTableProps extends Omit<BundleHeaderProps, 'searchDisabled'> {
+interface ReleasesTableProps extends Omit<ReleaseHeaderProps, 'searchDisabled'> {
   bundles: TableBundle[]
 }
 
-export function BundlesTable({bundles, searchTerm, setSearchTerm}: BundlesTableProps) {
+export function ReleasesTable({bundles, searchTerm, setSearchTerm}: ReleasesTableProps) {
   const tableContent = useMemo(() => {
     if (bundles.length === 0) {
       return (
@@ -49,12 +49,12 @@ export function BundlesTable({bundles, searchTerm, setSearchTerm}: BundlesTableP
       )
     }
 
-    return bundles.map((bundle) => <BundleRow key={bundle.name} bundle={bundle} />)
+    return bundles.map((bundle) => <ReleaseRow key={bundle.name} bundle={bundle} />)
   }, [bundles])
 
   return (
     <Stack as="table" space={1}>
-      <BundleHeader
+      <ReleaseHeader
         searchDisabled={!searchTerm && !bundles.length}
         searchTerm={searchTerm}
         setSearchTerm={setSearchTerm}
